chore(component): remove RoomTypeCard file
