--- src/app/(room)/room-type/_aggregation.tsx
@@ -1,18 +1,0 @@
-interface RoomTypeCardProps {
-  _id: string;
-  name: string;
-  description: string;
-  price: number;
-  imageUrl: string;
-  imageUrlList: string[];
-  areaInfo: string;
-  bedInfo: string;
-  maxPeople: number;
-}
-
-interface SquareCardProps {
-  title: string | number;
-  children: React.ReactNode;
-}
-
-export type { RoomTypeCardProps, SquareCardProps };
